test: harden macOS input policy
