+++ tests/policy/source-macos-input-policy.test.ts
@@ -1,27 +1,35 @@
 import assert from "node:assert/strict";
-import { readFile } from "node:fs/promises";
+import { readdir, readFile } from "node:fs/promises";
+import path from "node:path";
 import { describe, it } from "node:test";
 
-const productionInputSurface = [
-  "src/shared/contracts.ts",
-  "src/main/core/settings.ts",
-  "src/renderer/gw-native.d.ts",
-  "src/renderer/harness.ts",
-  "src/renderer/index.html",
-  "src/renderer/input.ts",
-  "src/renderer/settings.ts",
+const policyRoots = [
+  "src",
+  "docs/internals.md",
+  "docs/user-guide.md",
 ] as const;
+
+async function sourceFiles(entry: string): Promise<string[]> {
+  const children = await readdir(entry, { withFileTypes: true }).catch(() => null);
+  if (!children) return [entry];
+  return (
+    await Promise.all(
+      children.map((child) => sourceFiles(path.join(entry, child.name))),
+    )
+  ).flat();
+}
 
 describe("the macOS input policy", () => {
   it("has no selectable or persisted touch mode", async () => {
+    const files = (await Promise.all(policyRoots.map(sourceFiles))).flat();
     const sources = await Promise.all(
-      productionInputSurface.map((path) => readFile(path, "utf8")),
+      files.map(async (file) => [file, await readFile(file, "utf8")] as const),
     );
-    for (const [index, source] of sources.entries()) {
+    for (const [file, source] of sources) {
       assert.doesNotMatch(
         source,
         /touchMode|Mobile touch compatibility|Mouse and touch together|Translate mouse to touch/,
-        productionInputSurface[index],
+        file,
       );
     }
   });
